+++ CSharp/DevVmPowershell/Helpers.Tests.Integration/Tests/InstanceSettingsHelperTests.cs
@@ -21,7 +21,7 @@ namespace Helpers.Tests.Integration.Tests
 				sqlAdminUserName: TestConstants.SQL_USER_NAME,
 				sqlAdminPassword: TestConstants.SQL_PASSWORD);
 			IRestHelper restHelper = new RestHelper();
-			Sut = new InstanceSettingsHelper(connectionHelper, restHelper, TestConstants.RELATIVITY_INSTANCE_NAME, TestConstants.RELATIVITY_ADMIN_USER_NAME, TestConstants.RELATIVITY_ADMIN_PASSWORD);
+			Sut = new InstanceSettingsHelper(connectionHelper, restHelper);
 		}
 
 		[TearDown]